Adds the ability to add  comboboxes to armorUI
I have some clamping of the combo numbers from 2 to 25 because having  1 or 100 indexs is useless. To be able to modify the combobox I had to add a buffer to the p_id's because the comboboxes would of shared the same contents with the handles. Maybe I should do it in another way.