Support a default behaviour under Krom
The issue we have right now is if we don't specify the correct command to start armory2d under Krom we silently fail. This addresses this by making a similar behavior to what is done for debug_html5 when users don't specify a path to the canvas and makes i so we only need to specify the Krom folder or nothing if the exe is in the same path as the krom.js file.